Modify TTScrollView to pass the tapped location to TTScrollViewDelegate.
TTScrollView is designed to call delegate's scrollView:tapped: method when a user taps on the scroll view. Though the delegate does get called at the right time, the second argument of this call, (UITouch *)touch, actually become invalid when the delegate receives it. If you use the debugger to take a look at the content of this UITouch object within a scrollView:tapped: implementation, you'll find the "view" and "window" properties of this object are both set to nil.

According to SDK documentation for the UITouch class, "You should never retain an UITouch object when handling an event. If you need to keep information about a touch from one phase to another, you should copy that information from the UITouch object."

Because TTScrollView's tap detection requires a short delay to differentiate between single taps and double taps, TTScrollView schedules the delay with NSTimer and passes UITouch object to the callback via NSTimer's userInfo property. Because of this, after the delay finishes and when the UITouch object is restored from the userInfo property, you'll find that the "view" and "window" property are both set to nil.

This commit introduces a backward-compatible way to pass the touch point to a scroll view delegate. It does so by adding a new method to the TTScrollViewDelegate that, if implemented, allows the delegate to receive the point where the user touched the scroll view. If you don't implement the new method, the code will execute exactly like before.